requirements.txt
absl-py==0.9.0
cycler==0.10.0
fonttools==4.2.2
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.0
pandas==0.25.3
pkg-resources==0.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
regex==2019.12.20
six==1.13.0
